Cambios para variedad de Juego

+++ src/Personajes/Equipment/Front.java
@@ -2,9 +2,9 @@ package Personajes.Equipment;
 
 public class Front extends Equipment {
 
-    public float weight;
+    public double weight;
 
-    public Front(String name, String description, float weight, float defenseMultiplier) {
+    public Front(String name, String description, double weight, double defenseMultiplier) {
         super(name, description);
         this.weight = weight;
         this.multiplier = defenseMultiplier;
