+++ src/Personajes/Equipment/Helmet.java
@@ -2,7 +2,7 @@ package Personajes.Equipment;
 
 public class Helmet extends Equipment {
 
-    public Helmet(String name, String description, float defenseMultiplier) {
+    public Helmet(String name, String description, double defenseMultiplier) {
         super(name, description);
         this.multiplier = defenseMultiplier;
     }
