+++ src/Personajes/Equipment/Weapon.java
@@ -2,7 +2,7 @@ package Personajes.Equipment;
 
 public class Weapon extends Equipment {
 
-    public Weapon(String name, String description, float attackMultiplier) {
+    public Weapon(String name, String description, double attackMultiplier) {
         super(name, description);
         this.multiplier = attackMultiplier;
     }
